adding tests for keys support
